Fix linters errors
Change file names to snake_case
